Day 8
- Logic Gates system revamped
- First 4 levels done
- Level Ranking system
- Many Bugfixes and QOL | Code Cleanup
- Full OST loop

+++ Chroma Bounce/Assets/_Project/Scripts/UI/StepUIPrefab.cs
@@ -58,7 +58,7 @@ public class StepUIPrefab : MonoBehaviour{
     IEnumerator SetGunRotationI(string val){
         yield return new WaitForSecondsRealtime(0.05f);
         Debug.Log("Setting Gun rotation");
-        var i=int.Parse(val);
+        var i=float.Parse(val);
         stepProperties.gunRotation=i;
         SelectButton();
     }
